clean up aws
